Task 7 has been done

+++ product-service/src/product_by_id.py
@@ -52,7 +52,7 @@ def handler(event, context):
             'id': product['id'],
             'title': product['title'],
             'description': product['description'],
-            'price': int(product['price']),
+            'price': float(product['price']),
             'count': int(stock['count'])
         }
